requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
altair==5.5.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
comtypes==1.1.14
contourpy==1.0.5
cryptography==42.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.0.3
django-allauth==0.61.1
django-countries==7.6
django-crispy-forms==2.1
django-storages==1.14.2
dnspython==2.6.1
email_validator==2.2.0
entrypoints==0.4
executing==1.1.1
fastjsonschema==2.16.2
Flask==2.2.2
flatbuffers==24.3.25
fonttools==4.37.4
frozenlist==1.4.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.68.0
h5py==3.12.1
homepluscontrol==0.1
idna==3.4
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.21.0
jupyter_client==7.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==3.7.0
kiwisolver==1.4.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
ml-dtypes==0.4.1
MouseInfo==0.1.3
multidict==6.0.5
namex==0.0.8
narwhals==1.32.0
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.1
notebook-shim==0.1.0
numpy==1.26.4
oauthlib==3.2.2
opencage==2.3.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt_einsum==3.4.0
optree==0.13.1
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.6
phonenumbers==8.13.23
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.24.1
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==5.28.3
psutil==5.9.2
pure-eval==0.2.2
pyarrow==19.0.1
PyAudio==0.2.12
PyAutoGUI==0.9.53
pycparser==2.21
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.13.0
pyjokes==0.6.0
PyJWT==2.8.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
python-dateutil==2.8.2
python3-openid==3.2.0
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.4
pywhatkit==5.4
pywin32==304
pywinpty==2.0.8
PyYAML==6.0.2
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.9.4
scapy @ file:///D:/DataScience2023/HomeControl/scapy
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
Send2Trash==1.8.0
six==1.16.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
sqlparse==0.4.4
stack-data==0.5.1
statsmodels==0.14.1
streamlit==1.44.0
stripe==8.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
terminado==0.16.0
threadpoolctl==3.2.0
tinycss2==1.1.1
toml==0.10.2
tornado==6.2
traitlets==5.4.0
typing_extensions==4.13.0
tzdata==2023.3
urllib3==1.26.12
watchdog==6.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wikipedia==1.4.0
winspeech==1.0.1
wrapt==1.17.0
xgboost==3.0.0
yarl==1.9.4
